Added use_scene_labels and use_distortion_labels to PairDatasetArguments
Now you can choose to not include the scene and distortion question and
answer pairs from PairDataset to do ablation studies.

Removed labels_to_use because it's better to use two bools.

To provide context when training the MLLM to predict the quality levels,
we will have it predict the scene and distortion types of the images
before predicting the quality levels, in the same coversation. To do
this, distortion_type_message will have scene_type_message prepended to
it depending on use_scene_labels is true, and quality_message will have
scene_type_message and/or distortion_type_message prepended to it
depending on whether use_scene_labels and/or use_distortion_labels is
true.