Removed dead code in metrics.py
Added notebook to the dev dependencies so that a Jupyter Notebook server
can be launched in tmux, so that notebooks can continue to run when the
user disconnects from SSH.

Jupyter Notebooks have the problem of the outputs of cells not updating
after reconnecting to the server, so train_sequential_configurable.ipynb
now also logs its outputs to nb.log so that the user can continue
reading the outputs after reconnecting.

Added nb.log to .gitignore.

Added a cell in train_sequential_configurable.ipynb that runs the
evaluation script.

Fixed IQATrainer.compute_loss having an incompatible signature to the
base function.

Added some logging to IQATrainer.evaluation_loop that logs whether the
model is predicting similar quality scores for all images.